refactor(interactive-shell): split ui/streaming/__init__.py, retire paint vocabulary
Splits the 516-line ui/streaming/__init__.py into three named modules:
- renderer.py: Markdown block and response-header rendering.
- loop.py: the paragraph-by-paragraph streaming loop (stream_to_console,
  stream_to_console_state, publish_full_response, StreamRenderResult).
- closer.py: flushing a deferred Want-me-to closer after gather normalizes
  the answer.

__init__.py now contains only imports and __all__ (unchanged, same names
and same importable path), plus a re-export of rich.markdown.Markdown:
renderer._build_markdown_block constructs through the package façade rather
than a plain import so the existing test suite's
`streaming_module.Markdown = _SpyMarkdown` substitution keeps reaching the
real constructor call after the split.

Retires the "paint" vocabulary per the issue: StreamPaintResult ->
StreamRenderResult, _paint_paragraph_body -> _render_paragraph_body, and
the ShellOutputSink._paint field/local vars in agent_harness_adapters.py
(the type's other importer) -> _stream_result. grep -rn "Paint|_paint"
surfaces/ now returns nothing.

Closes #5445
